Changes to Variables and History.
VariablesView: renamed, double click opens HistoryDialog and button to show all fields of a variable
HistoryView: renamed, double click on root elements sets the step and double click on other elements opens their history, selection of this shows all histories on the root level.

+++ accidit-eclipse/src/de/hpi/accidit/eclipse/handlers/ShowVariableHistoryHandler.java
@@ -32,7 +32,7 @@ public class ShowVariableHistoryHandler extends AbstractHandler {
 		NamedValueNode nvn = (NamedValueNode) selectedLocals.getFirstElement();
 		NamedValue nv = (NamedValue) nvn.getValue();
 		int selectedNamedValueId = nv.getId();
-		long testId = TraceNavigatorUI.getGlobal().getTestId();
+		int testId = TraceNavigatorUI.getGlobal().getTestId();
 
 		NamedEntity[] options = null;
 		HistorySource src;
